+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log (@egomobile/api-utils)
 
-## 0.2.0
+## 0.3.0
+
+- add [extendRequest()](https://egomobile.github.io/node-api-utils/modules.html#extendRequest) middleware
+- add [handleApiValidationError()](https://egomobile.github.io/node-api-utils/modules.html#handleApiValidationError)
+- fix [ApiResponseData](https://egomobile.github.io/node-api-utils/modules.html#ApiResponseData)
+- set `sourceMap` to `(false)`
+- other (bug-)fixes
+
+## 0.2.1
 
 - add [parseListQuery](https://egomobile.github.io/node-api-utils/modules.html#parseListQuery), which creates middlewares to parse and validate query params, which can be used to realize list / collection cursors